requirements.txt
matplotlib==3.5.1
mtcnn==0.1.1
numpy==1.21.4
opencv_python==4.5.4.60
PyYAML==6.0
tensorflow==2.7.0
